+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v2.1.1 (2022-04-12)
+
+- Check if Declarative Shadow DOM is present before imperatively attaching to host element.
+
 ## v2.1.0 (2022-04-07)
 
 - Export the defined custom element as `web-share-defined.js` in case you don't want to manualy define it.